requirements.txt
# Pole Navigation Dependencies

# Core
numpy>=1.21.0
opencv-python>=4.5.0
pyyaml>=6.0

# ML
ultralytics>=8.0.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Optional (for advanced features)
# torch>=2.0.0  # Usually pre-installed on Jetson
# torchvision>=0.15.0
